[feat] bundles this script with setuptools
Modularizes the source code so that it can be installed
and run as a command line script, and adds missing dependencies.
