+++ .gitignore
@@ -1,2 +1,5 @@
 .idea/
 .python-version
+build/
+*.pyc
+*.egg-info/
