+++ requirements.txt
@@ -3,3 +3,4 @@ click
 clickhouse-connect
 psycopg2-binary
 pymongo[srv]
+requests
