+++ xapi_db_load/xapi/xapi_hint_answer.py
@@ -1,13 +1,15 @@
 from uuid import uuid4
 import json
 
-from xapi.xapi_common import XAPIBase
+from .xapi_common import XAPIBase
 
 
-# TODO: There are various other problem samples we should probably include eventually:
-# https://github.com/openedx/event-routing-backends/tree/master/event_routing_backends/processors/xapi/tests/fixtures/expected
-class BaseProblemCheck(XAPIBase):
-    problem_type = None  # "browser" or "server"
+class HintAnswerBase(XAPIBase):
+    verb_display = "asked"
+    verb = "http://adlnet.gov/expapi/verbs/asked"
+
+    # Whether this is a hint or an answer, "hint" or "answer" are valid values
+    type = None
 
     def get_data(self):
         event_id = str(uuid4())
@@ -17,7 +19,7 @@ class BaseProblemCheck(XAPIBase):
         emission_time = course.get_random_emission_time()
 
         e = self.get_randomized_event(
-            event_id, actor_id, course.course_url, problem_id, emission_time
+            event_id, actor_id, course, problem_id, emission_time
         )
 
         return {
@@ -25,56 +27,41 @@ class BaseProblemCheck(XAPIBase):
             "verb": self.verb,
             "actor_id": actor_id,
             "org": course.org,
-            "problem_id": problem_id,
             "course_run_id": course.course_url,
             "emission_time": emission_time,
             "event": e,
         }
 
-    def get_randomized_event(
-        self, event_id, account, course_locator, problem_id, create_time
-    ):
-        browser_object = {
+    def get_randomized_event(self, event_id, account, course, problem_id, create_time):
+        hint_object = {
             "object": {
                 "definition": {
-                    "type": "http://adlnet.gov/expapi/activities/cmi.interaction"
+                    "type": "https://w3id.org/xapi/acrossx/extensions/supplemental-info"
                 },
-                "id": problem_id,
+                "id": f"{problem_id}/hint/1",
                 "objectType": "Activity",
             }
         }
 
-        server_object = {
+        answer_object = {
             "object": {
-                "definition": {
-                    "extensions": {"http://id.tincanapi.com/extension/attempt-id": 10},
-                    "description": {
-                        "en-US": "Add the question text, or prompt, here. This text is required."
-                    },
-                    "interactionType": "other",
-                    "type": "http://adlnet.gov/expapi/activities/cmi.interaction",
-                },
-                "id": problem_id,
+                "definition": {"type": "http://id.tincanapi.com/activitytype/solution"},
+                "id": f"{problem_id}/answer",
                 "objectType": "Activity",
-            },
-            "result": {
-                "response": "A correct answer",
-                "score": {"max": 1, "min": 0, "raw": 0, "scaled": 0},
-                "success": False,
             },
         }
 
         event = {
             "id": event_id,
             "actor": {
-                "objectType": "Agent",
                 "account": {"homePage": "http://localhost:18000", "name": account},
+                "objectType": "Agent",
             },
             "context": {
                 "contextActivities": {
                     "parent": [
                         {
-                            "id": course_locator,
+                            "id": course.course_url,
                             "objectType": "Activity",
                             "definition": {
                                 "name": {"en-US": "Demonstration Course"},
@@ -92,21 +79,17 @@ class BaseProblemCheck(XAPIBase):
             "version": "1.0.3",
         }
 
-        if self.problem_type == "browser":
-            event.update(browser_object)
+        if self.type == "hint":
+            event.update(hint_object)
         else:
-            event.update(server_object)
+            event.update(answer_object)
 
         return json.dumps(event)
 
 
-class BrowserProblemCheck(BaseProblemCheck):
-    verb = "http://adlnet.gov/expapi/verbs/attempted"
-    verb_display = "attempted"
-    problem_type = "browser"
+class ShowHint(HintAnswerBase):
+    type = "hint"
 
 
-class ServerProblemCheck(BaseProblemCheck):
-    verb = "http://adlnet.gov/expapi/verbs/evaluated"
-    verb_display = "evaluated"
-    problem_type = "server"
+class ShowAnswer(HintAnswerBase):
+    type = "answer"
